requirements.txt
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1
django-filter==24.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
holehe==1.61
httpcore==1.0.5
httpx==0.27.2
idna==3.10
outcome==1.3.0.post0
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stripe==11.2.0
termcolor==2.4.0
tqdm==4.66.5
trio==0.26.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
